requirements.txt
torch
torchvision
segmentation-models-pytorch
python-telegram-bot
numpy
opencv-python-headless
